# docs/openapi/swagger.yaml
components:
  schemas:
...
    ? github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-github_com_Mininglamp-OSS_octo-marketplace_internal_api_response_EmptyResp
    : properties:
        data:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.EmptyResp'
      type: object
...
    ? github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-github_com_Mininglamp-OSS_octo-marketplace_internal_service_skill_SkillItem
    : properties:
        data:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_service_skill.SkillItem'
      type: object
...
    github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-internal_api_handler_skill_SkillMDResponse:
      properties:
        data:
          $ref: '#/components/schemas/internal_api_handler_skill.SkillMDResponse'
      type: object
...
    github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.EmptyResp:
      type: object
    github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error:
      properties:
        error:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.ErrorBody'
      type: object
    github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.ErrorBody:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        hint:
          type: string
        message:
          type: string
      type: object
...
    ? github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.OffsetList-github_com_Mininglamp-OSS_octo-marketplace_internal_service_skill_SkillItem
    : properties:
        data:
          items:
            $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_service_skill.SkillItem'
          type: array
          uniqueItems: false
        pagination:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.OffsetPagination'
      type: object
    github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.OffsetPagination:
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
      type: object
...
    internal_api_handler_skill.AdminCreateRequest:
      properties:
        category_id:
          type: string
        description:
          type: string
        display_name:
          type: string
        icon_url:
          type: string
        name:
          type: string
        parse_task_id:
          type: string
        tags:
          items:
            type: string
          type: array
          uniqueItems: false
        version:
          type: string
      required:
      - parse_task_id
      type: object
    internal_api_handler_skill.AdminReuploadRequest:
      properties:
        changelog:
          type: string
        parse_task_id:
          type: string
        tags:
          items:
            type: string
          type: array
          uniqueItems: false
        version:
          type: string
      required:
      - parse_task_id
      type: object
    internal_api_handler_skill.AdminUpdateRequest:
      properties:
        category_id:
          type: string
        description:
          type: string
        display_name:
          type: string
        icon_url:
          type: string
        name:
          type: string
        tags:
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
...
    internal_api_handler_skill.SkillMDResponse:
      properties:
        content:
          type: string
      type: object
...
  securitySchemes:
    Bearer:
      in: header
      name: Authorization
      type: apiKey
info:
  contact:
    name: OCTO API Team
    url: https://github.com/Mininglamp-OSS/octo-marketplace
  description: Skill and MCP marketplace API for OCTO.
  title: Octo Marketplace API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /admin/skills:
    get:
      description: List all visibility=public skills without Space restriction.
      operationId: admin_skill.list
      parameters:
      - description: Search name/description
        in: query
        name: q
        schema:
          type: string
      - description: Category ID
        in: query
        name: category_id
        schema:
          type: string
      - description: Comma-separated tags
        in: query
        name: tags
        schema:
          type: string
      - description: 'Sort: latest, downloads, views, comprehensive'
        in: query
        name: sort
        schema:
          type: string
      - description: Page number, default 1
        in: query
        name: page
        schema:
          type: integer
      - description: Page size
        in: query
        name: page_size
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.OffsetList-github_com_Mininglamp-OSS_octo-marketplace_internal_service_skill_SkillItem'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - Bearer: []
      summary: List public skills (admin)
      tags:
      - admin_skill
    post:
      description: Create a system skill with visibility=public.
      operationId: admin_skill.create
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_api_handler_skill.AdminCreateRequest'
                description: Skill
                summary: body
        description: Skill
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-github_com_Mininglamp-OSS_octo-marketplace_internal_service_skill_SkillItem'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: VALIDATION_ERROR
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: CONFLICT
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - Bearer: []
      summary: Create public skill (admin)
      tags:
      - admin_skill
  /admin/skills/{skill_id}:
    delete:
      description: Delete a public skill and its versions.
      operationId: admin_skill.delete
      parameters:
      - description: Skill ID
        in: path
        name: skill_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-github_com_Mininglamp-OSS_octo-marketplace_internal_api_response_EmptyResp'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - Bearer: []
      summary: Delete public skill (admin)
      tags:
      - admin_skill
    get:
      description: Return one public skill without Space restriction.
      operationId: admin_skill.get
      parameters:
      - description: Skill ID
        in: path
        name: skill_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-github_com_Mininglamp-OSS_octo-marketplace_internal_service_skill_SkillItem'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - Bearer: []
      summary: Get public skill detail (admin)
      tags:
      - admin_skill
    patch:
      description: Update basic info of a public skill.
      operationId: admin_skill.update
      parameters:
      - description: Skill ID
        in: path
        name: skill_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_api_handler_skill.AdminUpdateRequest'
                description: Skill changes
                summary: body
        description: Skill changes
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-github_com_Mininglamp-OSS_octo-marketplace_internal_service_skill_SkillItem'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: VALIDATION_ERROR
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: CONFLICT
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - Bearer: []
      summary: Update public skill (admin)
      tags:
      - admin_skill
...
  /admin/skills/{skill_id}/reupload:
    post:
      description: Upload a new zip to update the skill version.
      operationId: admin_skill.reupload
      parameters:
      - description: Skill ID
        in: path
        name: skill_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_api_handler_skill.AdminReuploadRequest'
                description: Reupload params
                summary: body
        description: Reupload params
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-github_com_Mininglamp-OSS_octo-marketplace_internal_service_skill_SkillItem'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: VALIDATION_ERROR
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: CONFLICT
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - Bearer: []
      summary: Reupload public skill version (admin)
      tags:
      - admin_skill
  /admin/skills/{skill_id}/skill_md:
    get:
      description: Return SKILL.md for a public skill without Space restriction.
      operationId: admin_skill.skillmd.get
      parameters:
      - description: Skill ID
        in: path
        name: skill_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-internal_api_handler_skill_SkillMDResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - Bearer: []
      summary: Get SKILL.md (admin)
      tags:
      - admin_skill
...
  /skills/{skill_id}/skill_md:
    get:
      description: Return the SKILL.md content for the current version of a visible
        Skill.
      operationId: skill.skillmd.get
      parameters:
      - description: Skill ID
        in: path
        name: skill_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-internal_api_handler_skill_SkillMDResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - Bearer: []
      summary: Get SKILL.md
      tags:
      - skill
...
servers:
- url: /v1
...